requirements.txt
absl-py==2.1.0
accelerate==0.27.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyascii==0.3.2
anyio==4.5.2
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
beautifulsoup4==4.12.2
bitarray==2.9.1
bitsandbytes==0.42.0
blessed==1.20.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
confection==0.1.3
contourpy==1.1.1
contractions==0.1.73
cycler==0.12.1
cymem==2.0.8
Cython==3.0.7
datasets==2.17.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
Distance==0.1.3
docker-pycreds==0.4.0
docstring_parser==0.16
dotmap==1.3.30
editdistance==0.6.2
einops==0.7.0
emoji==2.8.0
eval_type_backport==0.2.0
evaluate==0.4.1
exceptiongroup==1.2.2
executing==2.0.1
fairseq==0.12.2
fastapi==0.115.5
fastjsonschema==2.19.0
ffmpy==0.4.0
filelock==3.13.1
fonttools==4.44.3
frozenlist==1.4.0
fsspec==2023.10.0
ftfy==6.1.3
g2p-en==2.1.0
gitdb==4.0.11
GitPython==3.1.40
gpustat==1.1.1
gradio==4.44.1
gradio_client==1.3.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.21.4
hydra-core==1.0.7
idna==3.4
imageio==2.33.1
importlib-metadata==6.8.0
importlib-resources==6.1.1
imutils==0.5.4
inflect==7.0.0
ipykernel==6.26.0
ipython==8.12.3
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.10.0
llvmlite==0.41.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.0
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-ml-py==12.535.133
omegaconf==2.0.6
opencv-python-headless==4.8.1.78
orjson==3.10.11
packaging==23.2
pandas==2.0.3
parso==0.8.3
peft==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==4.0.0
plotly==5.18.0
pooch==1.8.0
portalocker==2.8.2
preshed==3.0.9
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyahocorasick==2.0.0
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydub==0.25.1
Pygments==2.16.1
pyparsing==3.1.1
pypinyin==0.50.0
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-multipart==0.0.17
pytorch-lightning==2.1.3
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
responses==0.18.0
rich==13.7.1
rouge-score==0.1.2
rpds-py==0.13.2
ruff==0.7.3
sacrebleu==2.4.0
sacremoses==0.1.1
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.35.0
setproctitle==1.3.3
shap==0.44.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
sklearn-hierarchical-classification==1.3.2
slicer==0.0.7
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.2
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
textsearch==0.0.24
thinc==8.2.1
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.12
tokenizers==0.15.2
tomlkit==0.12.0
torch==2.1.1+cu121
torchaudio==2.1.1+cu121
torchmetrics==1.2.1
torchvision==0.16.1+cu121
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.38.1
triton==2.1.0
trl==0.7.10
typer==0.13.0
typing_extensions==4.8.0
tyro==0.8.4
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.32.0
virtualenv-clone==0.5.7
wandb==0.16.0
wasabi==1.1.2
watermark==2.4.3
wcwidth==0.2.12
weasel==0.3.4
websockets==12.0
widgetsnbextension==4.0.10
xxhash==3.4.1
yarl==1.9.2
zhconv==1.4.3
zipp==3.17.0
